tests: extend test of dynamic detection to include extended emission
We want to make sure that the dynamic detection is robust to extended
emission (e.g., scattered light from bright stars). We do this by
inserting extra background over a large area. The dynamic detection
should be able to deal with this.

Unfortunately, some tuning is required for this extra background to get
a situation that is challenging but not impossible; details are included
in code comments.

Modified the configuration to match what is being used in
detectCoaddSources.py.